refactor(wizard): split integration health validators
