+++ app/cli/wizard/integration_validators/__init__.py
@@ -1,0 +1,1 @@
+"""Validator modules used by the onboarding integration health surface."""
